creates router package
