+++ templates/graphs/summary-graph.html
@@ -1,0 +1,8 @@
+{{ define "summary-graph"}}
+<div class="graph-container">
+    <canvas id="summary-graph"></canvas>
+</div>
+<script>
+    console.log("summary graph loaded")
+</script>
+{{end}}
