+++ templates/index.html
@@ -14,11 +14,14 @@
 
   <link rel="stylesheet" type="text/css" href="public/styles.css" />
   <script src="public/script.js"></script>
+  <script src="public/graphs.js"></script>
+
 </head>
 
 <body>
   <nav>
     <div class="nav-header">Bank Manager</div>
+    
     <div class="nav-menu">
       <button hx-get="/account" hx-target="#app" hx-swap="innerHTML">
         Account
